+++ CHANGELOG.md
@@ -68,3 +68,4 @@
 * Rename `ask` to `env` method of `Environmental` datatype
 * Introduce `><` type operator to separate functors from its arguments
 * Define `Determinable` typeclass and define its instance for `Predicate`
+* Define `curry` and `uncurry` for `Product` datatype